move report loading to reports
